Create Add todo test
